+++ ec/p55_lychrel_numbers.go
@@ -9,7 +9,7 @@ func P55() *problem {
 	return noInputNode(55, func(o command.Output) {
 		var count int
 		for i := 1; i < 10_000; i++ {
-			big := maths.NewInt(int64(i))
+			big := maths.NewInt(i)
 			big = big.Plus(big.Reverse())
 			for j := 0; j < 49; j++ {
 				if big.Palindrome() {
